web_editor/requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1
django-allauth==0.49.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-utils==0.0.2
djangorestframework==3.12.1
#djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==13.3.4
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
pycparser==2.21
#PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
gunicorn==20.1.0
djangorestframework-simplejwt==5.0.0
djangorestframework-simplejwt[crypto]